Partial cleanup of DFA and ATN
