+++ runtime/Rust/src/token_source.rs
@@ -88,47 +88,3 @@ where
         (**self).get_dfa_string()
     }
 }
-
-// / adaptor to feed parser with existing tokens
-// pub struct IterTokenSource<S, F> where S: Iterator, S::Item: Token, F: TokenFactory<Tok=S::Item> {
-//     iter: S,
-//     fact: F,
-// }
-//
-// impl<S, F> TokenSource for IterTokenSource<S, F> where S: Iterator, S::Item: Token, F: TokenFactory<Tok=S::Item> {
-//     type Tok = S::Item;
-//
-//     fn next_token(&mut self) -> Box<Self::Tok> {
-//         self.iter.next().map(Box::new).unwrap_or_else(
-//             || self.get_token_factory().create(
-//                 None,
-//                 EOF,
-//                 TOKEN_DEFAULT_CHANNEL,
-//                 -1,
-//                 -1,
-//                 self.get_line(),
-//                 self.get_char_position_in_line(),
-//             )
-//         )
-//     }
-//
-//     fn get_line(&self) -> isize {
-//         0
-//     }
-//
-//     fn get_char_position_in_line(&self) -> isize {
-//         -1
-//     }
-//
-//     fn get_input_stream(&mut self) -> Option<&mut dyn CharStream> {
-//         None
-//     }
-//
-//     fn get_source_name(&self) -> String {
-//         "<iterator>".to_string()
-//     }
-//
-//     fn get_token_factory(&self) -> &dyn TokenFactory<Tok=Self::Tok> {
-//         &self.fact
-//     }
-// }
