+++ runtime/Rust/templates
@@ -1,0 +1,1 @@
+../../tool/resources/org/antlr/v4/tool/templates/codegen/Rust/
