--- runtime/Rust/templates/Rust.stg
@@ -1,1 +1,0 @@
-../../../tool/resources/org/antlr/v4/tool/templates/codegen/Rust/Rust.stg
